requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.5
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-filter==22.1
django-reverse-admin==2.9.6
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
